@@ -12,7 +12,7 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^7.3",
+        "php": "^7.4|^8.0",
         "illuminate/support": "^7.0|^8.0"
     },
     "require-dev": {
