@@ -4,5 +4,5 @@ namespace ArtisanBuild\UI\View\Components;
 
 class Fab extends Fa
 {
-    public $family = 'b';
+    public string $family = 'b';
 }
