@@ -4,5 +4,5 @@ namespace ArtisanBuild\UI\View\Components;
 
 class Fad extends Fa
 {
-    public $family = 'd';
+    public string $family = 'd';
 }
