@@ -4,5 +4,5 @@ namespace ArtisanBuild\UI\View\Components;
 
 class Fal extends Fa
 {
-    public $family = 'l';
+    public string $family = 'l';
 }
