@@ -4,5 +4,5 @@ namespace ArtisanBuild\UI\View\Components;
 
 class Fas extends Fa
 {
-    public $family = 's';
+    public string $family = 's';
 }
